+++ app/src/main/java/com/example/sanket/contactbooksanket/MyClickListener.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 
 public interface MyClickListener {
     public void myOnClick(int position);
-    public void myOnClick(int position, int id, int imgUser, String firstname, String secondname, String userPhone);
+    public void myOnClick(int position, long id, int imgUser, String firstname, String secondname, String userPhone);
 
 
 }
